test(per-row-sweep): migrate popup_wrap_indent + virtual_line_bg_and_wrap + line_wrap_parity
Per #2058 step 5 batch 3.

Three small e2e files migrated to the semantic framework using
the rendered_rows extension:
- popup_wrap_indent (1 test)
- virtual_line_bg_and_wrap (2 tests)
- line_wrap_parity (3 tests)

All source e2e files deleted (faithful migration).
