--- crates/fresh-editor/tests/e2e/line_wrap_parity.rs
@@ -1,199 +1,0 @@
-//! Parity tests for the single-source-of-truth invariant.
-//!
-//! The `LineWrapCache` / `ViewLine`-based refactor (see
-//! `docs/internal/line-wrap-cache-plan.md`) aims to have scroll math,
-//! cursor positioning, and scrollbar thumb sizing all agree with the
-//! rendered content.  Earlier tests (sweep, consistency, perf)
-//! exercise scroll-math; these tests verify the remaining two
-//! surfaces:
-//!
-//!   * **Cursor-on-screen parity.**  For a word-wrapped buffer, the
-//!     character at the cursor's reported hardware position must be
-//!     the character the cursor logically points to.  If
-//!     `cursor_screen_position` disagreed with the renderer (e.g. by
-//!     using char-wrap while the renderer word-wraps), pressing
-//!     Down/Right would put the hardware cursor on a different cell
-//!     than the user expects.
-//!
-//!   * **Thumb-vs-content parity.**  The scrollbar's visual row count
-//!     (thumb sizing) must equal the total rendered visual rows.  If
-//!     they disagreed (as they did pre-refactor on word-wrapped
-//!     buffers), the thumb's size would lie about scrollable content.
-
-use crate::common::harness::EditorTestHarness;
-use crossterm::event::{KeyCode, KeyModifiers};
-use fresh::config::Config;
-
-fn config_with_wrap() -> Config {
-    let mut config = Config::default();
-    config.editor.line_wrap = true;
-    config
-}
-
-/// Realistic word-wrapped buffer — the kind of text where
-/// `wrap_line`'s char-wrap and the renderer's word-boundary wrap
-/// disagreed before the refactor.
-fn word_wrapped_buffer() -> String {
-    let para: String = (0..25)
-        .map(|i| format!("word{:02}", i))
-        .collect::<Vec<_>>()
-        .join(" ");
-    let mut lines = Vec::new();
-    for _ in 0..6 {
-        lines.push(para.clone());
-    }
-    lines.push("TAIL_MARKER_XYZ".to_string());
-    lines.join("\n")
-}
-
-/// Cursor-on-screen parity: place the cursor on a known character in
-/// a wrapped line, render, and assert that the rendered cell at the
-/// cursor's hardware position matches what we think the cursor is on.
-///
-/// If `cursor_screen_position` computes a different (row, col) than
-/// the renderer drew the cursor at, the cell-under-cursor will have
-/// the wrong character.  Before the refactor, char-wrap could put the
-/// cursor one row off on a word-wrapped line.
-#[test]
-fn cursor_hardware_position_matches_content_under_cursor() {
-    for &width in &[60u16, 80, 100] {
-        let mut harness =
-            EditorTestHarness::with_config(width, 20, config_with_wrap()).expect("harness");
-        let fixture = harness
-            .load_buffer_from_text(&word_wrapped_buffer())
-            .expect("load");
-        std::mem::forget(fixture);
-        harness.render().expect("render");
-
-        // Move to start, then some chars right — land on a known char.
-        harness
-            .send_key(KeyCode::Home, KeyModifiers::CONTROL)
-            .expect("ctrl+home");
-        // Go down one line (into the first wrapped paragraph).
-        harness
-            .send_key(KeyCode::Down, KeyModifiers::NONE)
-            .expect("down");
-        // Right a few chars to land mid-line.  45 keeps us near a
-        // wrap boundary on many widths — exactly the place the
-        // old char-wrap/word-wrap drift showed up.
-        for _ in 0..45 {
-            harness
-                .send_key(KeyCode::Right, KeyModifiers::NONE)
-                .expect("right");
-        }
-        harness.render().expect("render");
-
-        // The character logically at the cursor is determined by the
-        // buffer's cursor position; we can derive it via the editor's
-        // primary cursor byte.
-        let cursor_byte = harness.editor().active_cursors().primary().position;
-        let buffer_content = harness.get_buffer_content().expect("content");
-        let expected_char = buffer_content.as_bytes().get(cursor_byte).copied();
-
-        // Hardware cursor position — where the renderer placed the cursor.
-        let (hw_col, hw_row) = harness.screen_cursor_position();
-        let row_text = harness.get_screen_row(hw_row as usize);
-
-        // The character UNDER the cursor (the cell the renderer drew
-        // at that position) should match what we expect.
-        //
-        // The row text includes gutter characters ("  NN │ ") before
-        // the content.  Walk the row to find the hw_col-th visual
-        // column and check that cell.
-        let row_chars: Vec<char> = row_text.chars().collect();
-        let at_cursor = row_chars.get(hw_col as usize).copied();
-
-        let expected = expected_char.map(|b| b as char);
-        // If the cursor is at EOL (past the last char), the expected
-        // char may be `\n` or `None`; skip the strict comparison in
-        // that case — we still require `at_cursor` to be whitespace
-        // (a blank cell or gutter extension), not random content.
-        if let Some(exp) = expected {
-            if !exp.is_ascii_whitespace() && exp != '\n' {
-                assert_eq!(
-                    at_cursor,
-                    Some(exp),
-                    "[w={width}] cursor hardware position ({hw_col},{hw_row}) shows \
-                     {at_cursor:?} but the cursor byte ({cursor_byte}) is at {exp:?}.\n\
-                     Row: {row_text:?}",
-                );
-            }
-        }
-    }
-}
-
-/// Thumb-vs-content parity: the total visual rows reported by the
-/// scrollbar thumb sizing should equal the count of content rows the
-/// user can scroll through.  We test this indirectly by scrolling to
-/// the very end and checking the scrollbar indicates "at bottom"
-/// (thumb at/near bottom of track).
-#[test]
-fn scrollbar_thumb_reaches_bottom_on_word_wrapped_buffer() {
-    const WIDTH: u16 = 80;
-    const HEIGHT: u16 = 20;
-
-    let mut harness =
-        EditorTestHarness::with_config(WIDTH, HEIGHT, config_with_wrap()).expect("harness");
-    let fixture = harness
-        .load_buffer_from_text(&word_wrapped_buffer())
-        .expect("load");
-    std::mem::forget(fixture);
-    harness.render().expect("render");
-
-    // Drag the thumb to the bottom of the track.  If the thumb's
-    // reported total rows underreported the buffer (as pre-refactor
-    // wrap_line did vs. the renderer's word-aware wrap), the drag
-    // would stop short of showing the TAIL_MARKER.
-    let scrollbar_col = WIDTH - 1;
-    let (first, last) = harness.content_area_rows();
-    harness
-        .mouse_drag(scrollbar_col, first as u16, scrollbar_col, last as u16)
-        .expect("drag");
-    harness.render().expect("render");
-
-    let content: Vec<String> = (first..=last).map(|r| harness.get_screen_row(r)).collect();
-    let visible = content.iter().any(|row| row.contains("TAIL_MARKER_XYZ"));
-    assert!(
-        visible,
-        "after dragging the scrollbar thumb to the bottom of the track, \
-         the buffer's TAIL_MARKER should be visible. Content:\n{}",
-        content.join("\n"),
-    );
-}
-
-/// Thumb consistency across widths: the thumb's reported "total rows"
-/// should scale monotonically with how many rows the content actually
-/// wraps to.  Narrower terminals → more rows.  This doesn't directly
-/// read the thumb's numbers (those are behind `pub(super)` in the
-/// renderer), so we check it via the observable end-state: a narrow
-/// viewport takes more scrollbar-drag distance per visual row than a
-/// wide one.  In practice we just verify that at two different
-/// widths, the same buffer reaches its end-marker via drag-to-bottom,
-/// with no over- or under-scroll of the viewport.
-#[test]
-fn drag_to_bottom_reaches_end_at_multiple_widths() {
-    for &width in &[50u16, 70, 90, 120] {
-        let mut harness =
-            EditorTestHarness::with_config(width, 20, config_with_wrap()).expect("harness");
-        let fixture = harness
-            .load_buffer_from_text(&word_wrapped_buffer())
-            .expect("load");
-        std::mem::forget(fixture);
-        harness.render().expect("render");
-
-        let scrollbar_col = width - 1;
-        let (first, last) = harness.content_area_rows();
-        harness
-            .mouse_drag(scrollbar_col, first as u16, scrollbar_col, last as u16)
-            .expect("drag");
-        harness.render().expect("render");
-
-        let content: Vec<String> = (first..=last).map(|r| harness.get_screen_row(r)).collect();
-        let visible = content.iter().any(|row| row.contains("TAIL_MARKER_XYZ"));
-        assert!(
-            visible,
-            "[w={width}] thumb drag should reach TAIL_MARKER. Content:\n{}",
-            content.join("\n"),
-        );
-    }
-}
