+++ crates/fresh-editor/tests/semantic/mod.rs
@@ -34,6 +34,7 @@ pub mod migrated_issue_1147_wrapped_nav;
 pub mod migrated_issue_1574_viewport_drift;
 pub mod migrated_large_file_viewport;
 pub mod migrated_line_wrap_full_visibility;
+pub mod migrated_line_wrap_parity;
 pub mod migrated_macros;
 pub mod migrated_marker_roundtrip;
 pub mod migrated_recenter;
@@ -49,6 +50,7 @@ pub mod migrated_multicursor_extras;
 pub mod migrated_multicursor_full;
 pub mod migrated_paste;
 pub mod migrated_paste_extras;
+pub mod migrated_popup_wrap_indent;
 pub mod migrated_persistence;
 pub mod migrated_persistence_b;
 pub mod migrated_search_modal_flows;
@@ -66,6 +68,7 @@ pub mod migrated_undo_redo_full;
 pub mod migrated_undo_save_point;
 pub mod migrated_unicode;
 pub mod migrated_unicode_grapheme;
+pub mod migrated_virtual_line_bg_and_wrap;
 pub mod migrated_virtual_lines;
 pub mod migrated_workspace;
 pub mod migrated_workspace_extras;
